restructured repo to make compatible with rushstack heft
