+++ config/rush-project.json
@@ -1,0 +1,3 @@
+{
+  "projectOutputFolderNames": ["dist"]
+}
