+++ package.json
@@ -79,6 +79,9 @@
   },
   "devDependencies": {
     "@aws-cdk/assert": "1.116.0",
+    "@rushstack/eslint-config": "2.3.4",
+    "@rushstack/heft": "^0.36.1",
+    "@rushstack/heft-jest-plugin": "0.1.18",
     "@types/jest": "25.1.5",
     "@types/node": "~16.7.1",
     "@typescript-eslint/eslint-plugin": "~4.29.3",
@@ -93,33 +96,6 @@
     "prettier": "~2.3.2",
     "ts-jest": "25.4.0",
     "typescript": "3.9.10"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testRegex": "(/tests/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "lib/*.ts",
-      "!node_modules/**"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "globals": {
-      "ts-jest": {
-        "diagnostics": {
-          "warnOnly": true
-        }
-      },
-      "testEnvironment": "node"
-    }
   },
   "bundledDependencies": [
     "aws-sdk"
