+++ tsconfig.json
@@ -1,39 +1,24 @@
 {
+  "$schema": "http://json.schemastore.org/tsconfig",
+
   "compilerOptions": {
-    "alwaysStrict": true,
-    "charset": "utf8",
+    "outDir": "lib",
+    "rootDirs": ["src/"],
+
+    "forceConsistentCasingInFileNames": true,
     "declaration": true,
-    "experimentalDecorators": true,
-    "incremental": true,
-    "inlineSourceMap": true,
+    "sourceMap": true,
+    "declarationMap": true,
     "inlineSources": true,
-    "lib": [
-      "es2018"
-    ],
-    "module": "CommonJS",
-    "newLine": "lf",
-    "noEmitOnError": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "resolveJsonModule": true,
+    "experimentalDecorators": true,
     "strict": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true,
-    "stripInternal": false,
-    "target": "ES2018",
-    "composite": false,
-    "tsBuildInfoFile": "tsconfig.tsbuildinfo"
+    "esModuleInterop": true,
+    "types": ["node"],
+
+    "module": "commonjs",
+    "target": "es2017",
+    "lib": ["es2017"]
   },
-  "include": [
-    "**\\*.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "tests"
-  ],
-  "_generated_by_jsii_": "Generated by jsii - safe to delete, and ideally should be in .gitignore"
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules", "lib"]
 }
